pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metr-timehorizons-by-domain"
version = "0.1.0"
description = "Auto-updating cross-domain AI time horizon tracker"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Contributors" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "requests>=2.31",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[tool.setuptools]
packages = ["pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
